added functionality

+++ backend/main.py
@@ -4,11 +4,10 @@ from fastapi.middleware.cors import CORSMiddleware
 
 from model import Session, GetSessionModel
 from database import (
-    fetch_one_session,
     fetch_all_sessions,
     create_session,
     delete_session,
-    update_session, get_existing_object
+    update_session, get_existing_object, is_deleted
 )
 
 app = FastAPI()
@@ -34,9 +33,10 @@ async def read_root():
 
 @app.post("/session", response_model=Session)
 async def add_session(session: Session):
-    doc = await get_existing_object(session.muid)
-    if doc:
-        raise HTTPException(403, "A session with this muid already exists")
+    document = await get_existing_object(session.muid)
+    if document:
+        if is_deleted(document):
+            raise HTTPException(403, "This session has already been deleted")
     response = await create_session(session)
     if response:
         return response
@@ -45,31 +45,39 @@ async def add_session(session: Session):
 
 @app.get("/session/{session_muid}", response_model=GetSessionModel)
 async def get_sessions_by_id(session_muid: str):
-    response = await fetch_one_session(session_muid)
-    if response:
-        return response
-    raise HTTPException(404, f"There is no todo item with muid: {session_muid}")
+    document = await get_existing_object(session_muid)
+    if document:
+        if is_deleted(document):
+            raise HTTPException(403, "This session has already been deleted")
+        return document
+    raise HTTPException(404, f"There is no session item with muid: {session_muid}")
 
 
 @app.get("/sessions", response_model=List[GetSessionModel])
 async def get_sessions():
     response = await fetch_all_sessions()
-    if response:
-        return response
-    raise HTTPException(400, "Something went wrong / Bad Request")
+    return response
 
 
 @app.delete("/session/{session_muid}")
 async def delete_sessions(session_muid: str):
-    response = await delete_session(session_muid)
-    if response:
-        return "Successfully deleted todo"
-    raise HTTPException(404, f"There is no todo item with muid: {session_muid}")
+    document = await get_existing_object(session_muid)
+    if document:
+        if is_deleted(document):
+            raise HTTPException(403, "This session has already been deleted")
+        response = await delete_session(session_muid)
+        if response:
+            return "Successfully deleted session"
+    raise HTTPException(404, f"There is no session item with muid: {session_muid}")
 
 
-@app.put("/session/{session_muid}", response_model=Session)
+@app.put("/session/{session_muid}")
 async def update_sessions(session_muid: str, session: Session):
-    response = await update_session(session_muid, session)
-    if response:
-        return response
-    raise HTTPException(404, f"There is no todo item with muid: {session.muid}")
+    document = await get_existing_object(session_muid)
+    if document:
+        if is_deleted(document):
+            raise HTTPException(403, "This session has already been deleted")
+        succeeded = await update_session(document['_id'], session)
+        if succeeded:
+            return "Successfully updated session"
+    raise HTTPException(404, f"There is no session item with muid: {session_muid}")
